+++ src/components/property-page-widgets/PropertyPageHero/component.js
@@ -1,19 +1,13 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { upperCase } from 'lodash';
 
 import { withResponsive } from 'utils/with-responsive';
 import { Hero } from 'collections/Hero';
-import { HorizontalGutters } from 'layout/HorizontalGutters';
-import { FlexContainer } from 'layout/FlexContainer';
-import { Gallery } from 'media/Gallery';
-import { getGalleryHeadingMarkup } from 'utils/get-gallery-heading-markup';
-import { Button } from 'elements/Button';
-import { ICON_NAMES } from 'elements/Icon';
 import { VIEW_MORE_PICTURES } from 'utils/default-strings';
-import { Divider } from 'elements/Divider';
+import { CHECK_OUR_AVAILABILITY } from 'utils/default-strings';
 
 import { BOTTOM_OFFSET } from './constants';
+import { getGalleryMarkup } from './utils/getGalleryMarkup';
 
 /**
  * A homepage hero displays a hero with heading and a search bar on desktop screens.
@@ -44,14 +38,18 @@ const Component = ({
     sizes: backgroundImageSizes,
     srcSet: backgroundImageSrcSet,
     placeholderImageUrl,
+    imageWidth: backgroundImageHeight,
+    imageHeight: backgroundImageWidth,
   } = images[0];
 
   return (
     <Hero
       activeNavigationItemIndex={activeNavigationItemIndex}
+      backgroundImageHeight={backgroundImageHeight}
       backgroundImageSizes={backgroundImageSizes}
       backgroundImageSrcSet={backgroundImageSrcSet}
       backgroundImageUrl={backgroundImageUrl}
+      backgroundImageWidth={backgroundImageWidth}
       bottomOffset={BOTTOM_OFFSET}
       headerLogoSizes={headerLogoSizes}
       headerLogoSrc={headerLogoSrc}
@@ -69,25 +67,12 @@ const Component = ({
       searchBarOnChangeInput={searchBarOnChangeInput}
       searchBarOnSubmit={searchBarOnSubmit}
     >
-      <FlexContainer alignItems="flex-end">
-        <HorizontalGutters>
-          <Gallery
-            heading={getGalleryHeadingMarkup(propertyName, ratingNumber)}
-            images={images}
-            trigger={
-              <Button
-                icon={ICON_NAMES.PLACEHOLDER}
-                isCompact
-                isPositionedRight
-                isSecondary
-              >
-                {upperCase(secondaryButtonText)}
-              </Button>
-            }
-          />
-        </HorizontalGutters>
-      </FlexContainer>
-      <Divider />
+      {getGalleryMarkup(
+        images,
+        propertyName,
+        ratingNumber,
+        secondaryButtonText
+      )}
     </Hero>
   );
 };
@@ -103,7 +88,7 @@ Component.defaultProps = {
   propertyName: null,
   ratingNumber: null,
   searchBarGetIsDayBlocked: undefined,
-  searchBarModalHeadingText: undefined,
+  searchBarModalHeadingText: CHECK_OUR_AVAILABILITY,
   searchBarSearchButton: undefined,
   secondaryButtonText: VIEW_MORE_PICTURES,
   searchBarOnChangeInput: undefined,
@@ -149,12 +134,16 @@ Component.propTypes = {
     PropTypes.shape({
       /** Alternative text to show if the image can't be loaded by the browser */
       alternativeText: PropTypes.string,
+      /** The natural height of the image. */
+      imageHeight: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
       /** The label text for the when the image is not found. */
       imageNotFoundLabelText: PropTypes.string,
       /** Title of the image to show when hovering it on desktop browsers */
       imageTitle: PropTypes.string,
       /** URL pointing to the image to display. */
       imageUrl: PropTypes.string.isRequired,
+      /** The natural width of the image. */
+      imageWidth: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
       /** A visible label for the image. */
       label: PropTypes.string.isRequired,
       /** URL pointing to the placeholder image to display. */