+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+# [1.114.0](https://github.com/lodgify/lodgify-ui/compare/v1.113.0...v1.114.0) (2019-01-22)
+
+
+### Bug Fixes
+
+* **FullBleed:** pass imageHeight and imageWidth props to ResponsiveImage ([17f9202](https://github.com/lodgify/lodgify-ui/commit/17f9202))
+* **getImageMarkup:** pass correct arguments to getIsFluid ([bac5722](https://github.com/lodgify/lodgify-ui/commit/bac5722))
+* **getPlaceholderImageMarkup:** pass correct arguments to getIsFluid ([b1592c7](https://github.com/lodgify/lodgify-ui/commit/b1592c7))
+* **Hero:** pass imageHeight and imageWidth props to FullBleed ([5bf9fff](https://github.com/lodgify/lodgify-ui/commit/5bf9fff))
+* **HomepageHero:** pass background image height and width props to Hero ([2318d74](https://github.com/lodgify/lodgify-ui/commit/2318d74))
+* **PropertyPageHero:** pass image width and height props to Gallery and Hero ([22be55f](https://github.com/lodgify/lodgify-ui/commit/22be55f))
+* **ResponsiveImage:** prevent image from pushing placeholder out of view ([ad50f7b](https://github.com/lodgify/lodgify-ui/commit/ad50f7b))
+
+
+### Features
+
+* **getIsFluid:** add isFluid to condition ([d379d8e](https://github.com/lodgify/lodgify-ui/commit/d379d8e))
+
+# [1.113.0](https://github.com/lodgify/lodgify-ui/compare/v1.112.0...v1.113.0) (2019-01-21)
+
+
+### Features
+
+* **getGalleryMarkup:** implement util ([ff0fbec](https://github.com/lodgify/lodgify-ui/commit/ff0fbec))
+* **PropertyPageHero:** implement getGalleryMarkup util ([ba721b2](https://github.com/lodgify/lodgify-ui/commit/ba721b2))
+
 # [1.112.0](https://github.com/lodgify/lodgify-ui/compare/v1.111.0...v1.112.0) (2019-01-21)